fix: correct cutoff type annotation int -> float in from_atoms and empty
Agent-Logs-Url: https://github.com/nronne/agedi/sessions/7d95a6cb-eb71-4e6f-8805-10ae66f1fbdd

+++ src/agedi/data/atoms_graph.py
@@ -301,7 +301,7 @@ class AtomsGraph(Data):
     def from_atoms(
         cls,
         atoms: Atoms,
-        cutoff: int = 6.0,
+        cutoff: float = 6.0,
         skin: Optional[float] = None,
         max_neighbors: Optional[int] = None,
         dtype: torch.dtype = torch.float,
@@ -423,7 +423,7 @@ class AtomsGraph(Data):
         return cls(**kwargs)
 
     @classmethod
-    def empty(cls, cutoff: int = 6.0, skin: Optional[float] = None, max_neighbors: Optional[int] = None) -> "AtomsGraph":
+    def empty(cls, cutoff: float = 6.0, skin: Optional[float] = None, max_neighbors: Optional[int] = None) -> "AtomsGraph":
         """Create an empty graph.
 
         Parameters
